+++ app/__init__.py
@@ -6,7 +6,7 @@ Application factory.
 """
 from flask import Flask
 
-from . import cli, config, exceptions, extensions, models, routes
+from . import cli, config, exceptions, extensions, routes, shell
 
 __version__ = "0.1.0"
 
@@ -19,9 +19,9 @@ def create_app() -> Flask:
     """
     app = Flask(__name__)
     config.init_app(app)
-    models.init_app(app)
     extensions.init_app(app)
     routes.init_app(app)
     exceptions.init_app(app)
+    shell.init_app(app)
     cli.init_app(app)
     return app